Add materialTheme field to ThemeData
Commit 2e94931 automatically adds a Material theme to the widget tree.
However, this overrides any custom Material theme that might have been
specified above the widget tree. Many apps might still use a mix of
material and fluent_ui widgets. Now that a Material theme is always
automatically added, allow a custom Material theme to be specified as
part of fluent_ui.ThemeData with the optional materialTheme field.
If this field is not specified, the behavior is the same as before (it
will generate a Material theme with the appropriate brightness based
on the brightness of the fluent_ui theme).