TreeView uses ListView for vertical scrolling
* Behavior by default is the same as before, where the widget height will
expand to fill the content (shrinkWrap=true)
* Algo to build the list of items to add to the ListView is changed so
that the _visible field is still updated, but invisible items are not
added. This allows the itemExtent or prototypeItem property of the
ListView to be used to efficiently only build widgets for the items
currently in view. (Previously, invisible items were added to the list
but rendered as a 0 height item.)
* The example was updated to show how to use scrolling, and also that
the old behavior still works as well.