Add onItemExpandToggle and onExpandToggle; fix unexpected behavior of lazy loading
- Add global onItemExpandToggle callback to TreeView.
- Add onExpandToggle callback to TreeViewItem.
-> The global callback gets executed before(!) the item's callback.
- Add documentation.

These callbacks can be used to load child items lazily, which fixes the unexpected behavior of triggering
child loading on invoking an item. The loading of children now can be triggered, when the item gets expanded. Then a loading indicator is shown if the loading takes some time. When loading is complete, the loading indicator is removed and the item gets expanded.