Add NavigationView.paneBodyBuilder
Useful to override the widget used as the body of the navigation view in
conjunction with pane items. This can be useful if you are using
router-based navigation, and the body of the navigation pane should be
determined by the current route rather than just by the currently
selected pane.