Add TreeView.deselectParentWhenChildrenDeselected
Can be set to false so that a parent item can remain selected even when
all of its children has been deselected.

Also refactor code that changes the selection state in a multi-selection
mode TreeView, creating some helper methods that can be used by
application code. This makes it easier to programmatically change the
selection state of items in application logic.