paneBodyBuilder does its own state management
A common use case for paneBodyBuilder is to use an IndexedStack to
manage the state of the widgets displayed inside the navigation body.
An IndexedStack does not play nicely with the PageView.builder that it
was being wrapped in. Now, if a paneBodyBuilder is used, assume it is
responsible for its own state management.

Fixes #679